# nursery-inventory-system/nursery-inventory-system/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plant-issues:
    get:
      operationId: listPlantIssues
      tags: [plant-issues]
      summary: List all plant issues
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: projectId
          schema: { type: integer }
      responses:
        "200":
          description: List of plant issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlantIssue"

    post:
      operationId: createPlantIssue
      tags: [plant-issues]
      summary: Create a plant issue (issue stock to a project)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantIssueInput"
      responses:
        "201":
          description: Created plant issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantIssue"
        "400":
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /plant-issues/{id}:
    get:
      operationId: getPlantIssue
      tags: [plant-issues]
      summary: Get a plant issue
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Plant issue record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantIssue"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Material Masters ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Plant Master schemas ─────────────────────────────────────────────────
...
    ProductionBatchPlant:
      type: object
      required: [id, plantCode, commonName, botanicalName, potSize]
      properties:
        id:            { type: integer }
        plantCode:     { type: string }
        commonName:    { type: string }
        botanicalName: { type: string }
        potSize:       { type: string }
        imageUrl:      { type: ["string", "null"] }
...
    PlantIssueProject:
      type: object
      required: [id, projectCode, projectName, clientName]
      properties:
        id:          { type: integer }
        projectCode: { type: string }
        projectName: { type: string }
        clientName:  { type: string }
    PlantIssueLine:
      type: object
      required:
        - id
        - plantId
        - plant
        - batchNumber
        - batchSource
        - potSize
        - issueQuantity
        - costPerPlant
        - totalCost
      properties:
        id:            { type: integer }
        plantId:       { type: integer }
        plant:
          $ref: "#/components/schemas/ProductionBatchPlant"
        batchNumber:   { type: string }
        batchSource:   { type: string, enum: [PURCHASE, PRODUCTION] }
        potSize:       { type: string }
        issueQuantity: { type: integer }
        costPerPlant:  { type: number }
        totalCost:     { type: number }
    PlantIssue:
      type: object
      required:
        - id
        - issueNumber
        - issueDate
        - projectId
        - project
        - requestedBy
        - issuedBy
        - totalQuantity
        - totalValue
        - totalMaterialQuantity
        - totalMaterialValue
        - lines
        - materialLines
        - createdAt
        - updatedAt
      properties:
        id:                    { type: integer }
        issueNumber:           { type: string }
        issueDate:             { type: string, format: date }
        projectId:             { type: integer }
        project:
          $ref: "#/components/schemas/PlantIssueProject"
        requestedBy:           { type: string }
        issuedBy:              { type: string }
        remarks:               { type: ["string", "null"] }
        totalQuantity:         { type: integer }
        totalValue:            { type: number }
        totalMaterialQuantity: { type: number }
        totalMaterialValue:    { type: number }
        lines:
          type: array
          items:
            $ref: "#/components/schemas/PlantIssueLine"
        materialLines:
          type: array
          items:
            $ref: "#/components/schemas/MaterialIssueLine"
        createdAt:             { type: string, format: date-time }
        updatedAt:             { type: string, format: date-time }
    PlantIssueLineInput:
      type: object
      required: [plantId, batchNumber, batchSource, potSize, issueQuantity]
      properties:
        plantId:       { type: integer }
        batchNumber:   { type: string, minLength: 1 }
        batchSource:   { type: string, enum: [PURCHASE, PRODUCTION] }
        potSize:       { type: string, minLength: 1 }
        issueQuantity: { type: integer, minimum: 1 }
    PlantIssueInput:
      type: object
      required: [issueDate, projectId, requestedBy, issuedBy, lines]
      properties:
        issueNumber: { type: string }
        issueDate:   { type: string, format: date }
        projectId:   { type: integer }
        requestedBy: { type: string, minLength: 1 }
        issuedBy:    { type: string, minLength: 1 }
        remarks:     { type: string }
        lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PlantIssueLineInput"
        materialLines:
          type: array
          items:
            $ref: "#/components/schemas/MaterialIssueLineInput"

    # ── Auth schemas ─────────────────────────────────────────────────────────
...
    MaterialMasterRef:
      type: object
      required: [id, materialCode, name, category, unit]
      properties:
        id:           { type: integer }
        materialCode: { type: string }
        name:         { type: string }
        category:     { type: string }
        unit:         { type: string }
...
    MaterialIssueLineInput:
      type: object
      required: [materialId, purchaseNumber, stockLocation, unit, issueQuantity]
      properties:
        materialId:     { type: integer }
        purchaseNumber: { type: string, minLength: 1 }
        stockLocation:  { type: string, minLength: 1 }
        unit:           { type: string, minLength: 1 }
        issueQuantity:  { type: number, minimum: 0.001 }
    MaterialIssueLine:
      type: object
      required:
        - id
        - materialId
        - material
        - purchaseNumber
        - stockLocation
        - unit
        - issueQuantity
        - unitCost
        - totalCost
      properties:
        id:             { type: integer }
        materialId:     { type: integer }
        material:
          $ref: "#/components/schemas/MaterialMasterRef"
        purchaseNumber: { type: string }
        stockLocation:  { type: string }
        unit:           { type: string }
        issueQuantity:  { type: number }
        unitCost:       { type: number }
        totalCost:      { type: number }

    # ── Dashboard schemas ────────────────────────────────────────────────────
...